backup: bug fix for improper struct dirent usage
Between Minix 3.1.2 and Minix 3.1.3, the d_name member of struct
dirent was reduced from 61 characters to 1 character. backup(8)
blindly copies the result of readdir() into an array of dirent
structs. As a consequence, the structs in the array only contain
the first letter of the name and the name string was not null
terminated. This caused the backup utility to always fail.

Remedy the situation by adding room for the full d_name.

Change-Id: I2a7b4630f1bb727825a95a6095a7a543fc2f08f6